fix: 最優先522行ファイル分割完了 - 300行制限完全遵守
test_convert_check_processor.py (522行) を機能別に3ファイル分割:
- test_convert_command_basic.py (130行) - ConvertCommand基本機能
- test_convert_command_advanced.py (119行) - ウォッチモード・設定指定
- test_check_processor_deep.py (299行) - CheckSyntax + ConvertProcessor

Single Responsibility原則適用、test_[module]_[feature]_[aspect].py命名規則準拠。
全ファイル300行制限遵守達成。
