+++ tests/test_convert_check_processor.py
@@ -1,522 +1,6 @@
-"""Convert Command and Check Syntax Deep Coverage Tests
-
-深度カバレッジ向上のためのCommands系モジュールテスト。
-target: 0%カバレッジのcommands系モジュールを大幅改善。
-"""
-
-import tempfile
-from pathlib import Path
-from unittest.mock import MagicMock, Mock, patch
-
-import pytest
-
-
-class TestConvertCommandDeep:
-    """ConvertCommand深度テスト - 84行 0%→80%+目標"""
-
-    def test_convert_command_initialization(self):
-        """ConvertCommand初期化テスト"""
-        from kumihan_formatter.commands.convert.convert_command import ConvertCommand
-
-        with (
-            patch(
-                "kumihan_formatter.commands.convert.convert_command.get_logger"
-            ) as mock_logger,
-            patch(
-                "kumihan_formatter.commands.convert.convert_command.get_console_ui"
-            ) as mock_ui,
-        ):
-
-            command = ConvertCommand()
-
-            # 初期化確認
-            assert command is not None
-            assert hasattr(command, "validator")
-            assert hasattr(command, "processor")
-            assert hasattr(command, "watcher")
-            assert hasattr(command, "friendly_error_handler")
-
-            # ログ初期化確認
-            mock_logger.assert_called_once()
-
-    def test_convert_command_execute_basic(self):
-        """基本的な変換実行テスト"""
-        from kumihan_formatter.commands.convert.convert_command import ConvertCommand
-
-        with (
-            patch(
-                "kumihan_formatter.commands.convert.convert_command.get_logger"
-            ) as mock_logger,
-            patch(
-                "kumihan_formatter.commands.convert.convert_command.get_console_ui"
-            ) as mock_ui,
-            patch("sys.exit") as mock_exit,
-        ):
-
-            command = ConvertCommand()
-
-            # モックで依存関係をコントロール
-            command.validator = Mock()
-            command.processor = Mock()
-            command.watcher = Mock()
-            command.friendly_error_handler = Mock()
-
-            # 構文チェックの結果をモック
-            command.validator.perform_syntax_check.return_value.to_console_output.return_value = (
-                "Syntax check passed"
-            )
-            command.validator.perform_syntax_check.return_value.has_errors.return_value = (
-                False
-            )
-
-            # 基本実行テスト
-            command.execute(
-                input_file="test.txt",
-                output="output/",
-                no_preview=False,
-                watch=False,
-                config=None,
-                show_test_cases=False,
-                template_name=None,
-                include_source=False,
-                syntax_check=True,
-            )
-
-            # バリデータが呼ばれたことを確認
-            command.validator.perform_syntax_check.assert_called_once()
-
-    def test_convert_command_execute_with_watch(self):
-        """ウォッチモード実行テスト"""
-        from kumihan_formatter.commands.convert.convert_command import ConvertCommand
-
-        with (
-            patch(
-                "kumihan_formatter.commands.convert.convert_command.get_logger"
-            ) as mock_logger,
-            patch(
-                "kumihan_formatter.commands.convert.convert_command.get_console_ui"
-            ) as mock_ui,
-        ):
-
-            command = ConvertCommand()
-            command.watcher = Mock()
-            command.validator = Mock()
-            command.processor = Mock()
-            command.friendly_error_handler = Mock()
-
-            # 構文チェック結果を適切にモック
-            mock_error_report = Mock()
-            mock_error_report.has_errors.return_value = False
-            mock_error_report.has_warnings.return_value = False
-            mock_error_report.to_console_output.return_value = "No errors"
-            command.validator.perform_syntax_check.return_value = mock_error_report
-            command.validator.validate_input_file.return_value = Path("test.txt")
-            command.validator.check_file_size.return_value = True
-
-            try:
-                command.execute(
-                    input_file="test.txt",
-                    output="output/",
-                    no_preview=True,
-                    watch=True,  # ウォッチモード
-                    config=None,
-                    show_test_cases=False,
-                    template_name=None,
-                    include_source=False,
-                    syntax_check=True,
-                )
-
-                # ウォッチャーが呼ばれることを期待（正しいメソッド名を使用）
-                command.watcher.start_watch_mode.assert_called_once()
-
-            except (AttributeError, SystemExit):
-                # モック設定の問題やSystemExitは許容
-                pass
-
-    def test_convert_command_error_handling(self):
-        """エラーハンドリングテスト"""
-        from kumihan_formatter.commands.convert.convert_command import ConvertCommand
-
-        with (
-            patch(
-                "kumihan_formatter.commands.convert.convert_command.get_logger"
-            ) as mock_logger,
-            patch(
-                "kumihan_formatter.commands.convert.convert_command.get_console_ui"
-            ) as mock_ui,
-        ):
-
-            command = ConvertCommand()
-
-            # エラーを発生させるモック
-            command.validator = Mock()
-            command.processor = Mock()
-            command.friendly_error_handler = Mock()
-
-            # ファイル検証でFileNotFoundErrorを発生させる
-            command.validator.validate_input_file.side_effect = FileNotFoundError(
-                "File not found"
-            )
-
-            try:
-                command.execute(
-                    input_file="nonexistent.txt",
-                    output="output/",
-                    no_preview=True,
-                    watch=False,
-                    config=None,
-                    show_test_cases=False,
-                    template_name=None,
-                    include_source=False,
-                    syntax_check=True,
-                )
-            except SystemExit:
-                # sys.exit(1)が呼ばれることを確認
-                pass
-            except Exception:
-                # その他のエラーハンドリングが動作することを確認
-                pass
-
-    def test_convert_command_with_config(self):
-        """設定ファイル指定テスト"""
-        from kumihan_formatter.commands.convert.convert_command import ConvertCommand
-
-        with (
-            patch(
-                "kumihan_formatter.commands.convert.convert_command.get_logger"
-            ) as mock_logger,
-            patch(
-                "kumihan_formatter.commands.convert.convert_command.get_console_ui"
-            ) as mock_ui,
-        ):
-
-            command = ConvertCommand()
-            command.validator = Mock()
-            command.processor = Mock()
-            command.friendly_error_handler = Mock()
-
-            # 構文チェック結果を適切にモック
-            mock_error_report = Mock()
-            mock_error_report.has_errors.return_value = False
-            mock_error_report.has_warnings.return_value = False
-            mock_error_report.to_console_output.return_value = "No errors"
-            command.validator.perform_syntax_check.return_value = mock_error_report
-            command.validator.validate_input_file.return_value = Path("test.txt")
-            command.validator.check_file_size.return_value = True
-            command.processor.convert_file.return_value = Path("output.html")
-
-            # 設定ファイル指定でのテスト
-            with tempfile.NamedTemporaryFile(
-                mode="w", suffix=".json", delete=False
-            ) as f:
-                f.write('{"template": "default"}')
-                config_path = f.name
-
-            try:
-                command.execute(
-                    input_file="test.txt",
-                    output="output/",
-                    no_preview=True,
-                    watch=False,
-                    config=config_path,  # 設定ファイル指定
-                    show_test_cases=False,
-                    template_name=None,
-                    include_source=False,
-                    syntax_check=True,
-                )
-            except SystemExit:
-                # 正常終了時のsys.exit(0)は許容
-                pass
-            except Exception:
-                # 依存関係の問題は許容
-                pass
-            finally:
-                Path(config_path).unlink(missing_ok=True)
-
-
-class TestCheckSyntaxDeep:
-    """CheckSyntax深度テスト - 73行 0%→70%+目標"""
-
-    def test_check_syntax_command_basic(self):
-        """基本的な構文チェックテスト"""
-        from kumihan_formatter.commands.check_syntax import CheckSyntaxCommand
-
-        try:
-            command = CheckSyntaxCommand()
-            assert command is not None
-
-            # 基本的なインターフェース確認
-            assert hasattr(command, "execute") or hasattr(command, "check")
-
-        except ImportError:
-            # モジュールが存在しない場合はスキップ
-            pytest.skip("CheckSyntaxCommand module not available")
-
-    def test_check_syntax_valid_file(self):
-        """有効ファイルの構文チェック"""
-        from kumihan_formatter.commands.check_syntax import CheckSyntaxCommand
-
-        try:
-            command = CheckSyntaxCommand()
-
-            # 有効なKumihan記法でテストファイル作成
-            valid_content = """# Test Document
-
-This is a test document with valid syntax.
-
-;;;highlight;;;
-Valid highlight block
-;;;
-
-Text with ((valid footnote)) notation.
-
-｜日本語《にほんご》 valid ruby notation.
-"""
-
-            with tempfile.NamedTemporaryFile(
-                mode="w", suffix=".txt", delete=False
-            ) as f:
-                f.write(valid_content)
-                test_file = f.name
-
-            try:
-                # 構文チェック実行（filesパラメータはリストである必要がある）
-                result = command.execute([test_file])
-
-                # 結果確認（エラーなしまたは有効な結果）
-                assert result is not None or result == 0
-
-            except AttributeError:
-                # メソッド名が違う場合の代替テスト
-                try:
-                    result = command.check([test_file])
-                    assert result is not None
-                except:
-                    # 基本的なオブジェクト作成確認のみ
-                    pass
-            except SystemExit as e:
-                # 正常終了のsys.exit(0)は許容
-                assert e.code == 0 or e.code is None
-            finally:
-                Path(test_file).unlink(missing_ok=True)
-
-        except ImportError:
-            pytest.skip("CheckSyntaxCommand module not available")
-
-    def test_check_syntax_invalid_file(self):
-        """無効ファイルの構文チェック"""
-        from kumihan_formatter.commands.check_syntax import CheckSyntaxCommand
-
-        try:
-            command = CheckSyntaxCommand()
-
-            # 無効なKumihan記法でテストファイル作成
-            invalid_content = """# Test Document
-
-;;;incomplete block without closing
-
-((incomplete footnote
-
-｜incomplete《ruby notation
-
-Invalid syntax patterns.
-"""
-
-            with tempfile.NamedTemporaryFile(
-                mode="w", suffix=".txt", delete=False
-            ) as f:
-                f.write(invalid_content)
-                test_file = f.name
-
-            try:
-                # 構文チェック実行（エラーを検出することを期待）
-                result = command.execute([test_file])
-
-                # エラー検出の確認（結果に応じて）
-                if result is not None:
-                    # エラーが検出された場合の結果確認
-                    assert isinstance(result, (int, bool, list, dict))
-
-            except AttributeError:
-                # メソッド名違いの代替
-                try:
-                    result = command.check([test_file])
-                except:
-                    pass
-            except SystemExit as e:
-                # 構文エラー検出による終了コード1は期待される
-                assert e.code == 1
-            except Exception as e:
-                # 構文エラー検出による例外は正常
-                assert "syntax" in str(e).lower() or "invalid" in str(e).lower()
-            finally:
-                Path(test_file).unlink(missing_ok=True)
-
-        except ImportError:
-            pytest.skip("CheckSyntaxCommand module not available")
-
-    def test_check_syntax_nonexistent_file(self):
-        """存在しないファイルの構文チェック"""
-        from kumihan_formatter.commands.check_syntax import CheckSyntaxCommand
-
-        try:
-            command = CheckSyntaxCommand()
-
-            # 存在しないファイルでテスト
-            nonexistent_file = "nonexistent_file.txt"
-
-            try:
-                result = command.execute([nonexistent_file])  # リスト形式で渡す
-                # ファイルエラーの適切な処理確認
-            except (FileNotFoundError, SystemExit):
-                # 期待される例外（SystemExitも含む）
-                pass
-            except AttributeError:
-                # メソッド名違いの場合
-                try:
-                    command.check([nonexistent_file])  # リスト形式で渡す
-                except FileNotFoundError:
-                    pass
-
-        except ImportError:
-            pytest.skip("CheckSyntaxCommand module not available")
-
-
-class TestConvertProcessorDeep:
-    """ConvertProcessor深度テスト - 119行 0%→80%+目標"""
-
-    def test_convert_processor_initialization(self):
-        """ConvertProcessor初期化テスト"""
-        from kumihan_formatter.commands.convert.convert_processor import (
-            ConvertProcessor,
-        )
-
-        with patch(
-            "kumihan_formatter.commands.convert.convert_processor.get_logger"
-        ) as mock_logger:
-            processor = ConvertProcessor()
-
-            assert processor is not None
-            # 基本属性の確認
-            assert (
-                hasattr(processor, "convert_file")
-                or hasattr(processor, "process")
-                or hasattr(processor, "convert")
-            )
-
-    def test_convert_processor_basic_processing(self):
-        """基本的な変換処理テスト"""
-        from kumihan_formatter.commands.convert.convert_processor import (
-            ConvertProcessor,
-        )
-
-        with patch(
-            "kumihan_formatter.commands.convert.convert_processor.get_logger"
-        ) as mock_logger:
-            processor = ConvertProcessor()
-
-            # テストファイル作成
-            test_content = """# Test Document
-
-This is a test document for conversion.
-
-;;;highlight;;; Important content ;;;
-
-Regular paragraph text.
-"""
-
-            with tempfile.NamedTemporaryFile(
-                mode="w", suffix=".txt", delete=False
-            ) as input_f:
-                input_f.write(test_content)
-                input_file = input_f.name
-
-            with tempfile.NamedTemporaryFile(
-                mode="w", suffix=".html", delete=False
-            ) as output_f:
-                output_file = output_f.name
-
-            try:
-                # 変換処理実行
-                if hasattr(processor, "process"):
-                    result = processor.process(input_file, output_file)
-                elif hasattr(processor, "convert"):
-                    result = processor.convert(input_file, output_file)
-                else:
-                    # メソッド名を推測してテスト
-                    methods = [
-                        attr for attr in dir(processor) if not attr.startswith("_")
-                    ]
-                    if methods:
-                        method = getattr(processor, methods[0])
-                        if callable(method):
-                            result = method(input_file, output_file)
-
-                # 結果確認
-                assert Path(output_file).exists() or result is not None
-
-            except Exception as e:
-                # 依存関係の問題は許容
-                assert "Mock" not in str(e) or "import" in str(e).lower()
-            finally:
-                Path(input_file).unlink(missing_ok=True)
-                Path(output_file).unlink(missing_ok=True)
-
-    def test_convert_processor_with_template(self):
-        """テンプレート指定変換テスト"""
-        from kumihan_formatter.commands.convert.convert_processor import (
-            ConvertProcessor,
-        )
-
-        with patch(
-            "kumihan_formatter.commands.convert.convert_processor.get_logger"
-        ) as mock_logger:
-            processor = ConvertProcessor()
-
-            # テンプレート指定での変換テスト
-            test_content = "# Simple Test"
-
-            with tempfile.NamedTemporaryFile(
-                mode="w", suffix=".txt", delete=False
-            ) as f:
-                f.write(test_content)
-                test_file = f.name
-
-            try:
-                # テンプレート指定でのテスト（メソッド存在確認）
-                if hasattr(processor, "set_template"):
-                    processor.set_template("default")
-
-                if hasattr(processor, "process"):
-                    result = processor.process(test_file, "output.html")
-
-            except Exception:
-                # 依存関係エラーは許容
-                pass
-            finally:
-                Path(test_file).unlink(missing_ok=True)
-
-    def test_convert_processor_error_handling(self):
-        """ConvertProcessorエラーハンドリングテスト"""
-        from kumihan_formatter.commands.convert.convert_processor import (
-            ConvertProcessor,
-        )
-
-        with patch(
-            "kumihan_formatter.commands.convert.convert_processor.get_logger"
-        ) as mock_logger:
-            processor = ConvertProcessor()
-
-            # 無効なファイルでのエラーテスト
-            try:
-                if hasattr(processor, "process"):
-                    processor.process("nonexistent.txt", "output.html")
-                elif hasattr(processor, "convert"):
-                    processor.convert("nonexistent.txt", "output.html")
-
-            except (FileNotFoundError, IOError):
-                # 期待されるエラー
-                pass
-            except Exception:
-                # その他のエラーも許容（依存関係問題など）
-                pass
+# 旧test_convert_check_processor.pyファイルは以下3ファイルに分割:
+# - test_convert_command_basic.py (ConvertCommand基本機能)
+# - test_convert_command_advanced.py (ConvertCommand高度機能)
+# - test_check_processor_deep.py (CheckSyntax + ConvertProcessor)
+#
+# この分割により522行→各174行以下に削減し、300行制限を遵守。
